# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChartPoint {
  d: string;
  p: number;
}

export interface Stock {
  ticker: string;
  company: string;
  price: number;
  exchange: string;
  industry: string;
  daysUnderOne: number;
  daysToDeadline: number;
  bounceProbability: number;
  delistingRisk: number;
  complianceScore: number;
  financialScore: number;
  operatorScore: number;
  industryScore: number;
  patternScore: number;
  tradabilityScore: number;
  operatorNote: string;
  financialNote: string;
  tradeWindow: string;
  entryZone: string;
  targetZone: string;
  stopZone: string;
  status: string;
  volume: string;
  chart: ChartPoint[];
  /** Model confidence 0–1 based on data freshness and source availability */
  confidenceScore: number;
  /** Unix epoch ms of the last successful live price fetch; 0 means static data only */
  dataFreshnessMs: number;
  /** Number of verified data sources contributing to this record */
  sourcesCount: number;
}

export interface WatchlistResponse {
  tickers: string[];
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  info: "info",
  warning: "warning",
  critical: "critical",
} as const;

export interface Alert {
  id: string;
  message: string;
  ticker: string;
  severity: AlertSeverity;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListStocksParams = {
  /**
   * Filter by status
   */
  status?: string;
  /**
   * Search query for ticker or company
   */
  q?: string;
};

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};