# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stocks:
    get:
      operationId: listStocks
      tags: [stocks]
      summary: List distressed stocks
      description: Returns all tracked distressed stocks with scoring data
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by status
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search query for ticker or company
      responses:
        "200":
          description: List of stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stock"
  /stocks/{ticker}:
    get:
      operationId: getStock
      tags: [stocks]
      summary: Get stock detail
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stock"
        "404":
          description: Stock not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ChartPoint:
      type: object
      properties:
        d:
          type: string
        p:
          type: number
      required:
        - d
        - p
    Stock:
      type: object
      properties:
        ticker:
          type: string
        company:
          type: string
        price:
          type: number
        exchange:
          type: string
        industry:
          type: string
        daysUnderOne:
          type: integer
        daysToDeadline:
          type: integer
        bounceProbability:
          type: integer
        delistingRisk:
          type: integer
        complianceScore:
          type: integer
        financialScore:
          type: integer
        operatorScore:
          type: integer
        industryScore:
          type: integer
        patternScore:
          type: integer
        tradabilityScore:
          type: integer
        operatorNote:
          type: string
        financialNote:
          type: string
        tradeWindow:
          type: string
        entryZone:
          type: string
        targetZone:
          type: string
        stopZone:
          type: string
        status:
          type: string
        volume:
          type: string
        chart:
          type: array
          items:
            $ref: "#/components/schemas/ChartPoint"
        confidenceScore:
          type: number
          description: Model confidence 0–1 based on data freshness and source availability
        dataFreshnessMs:
          type: integer
          description: Unix epoch ms of the last successful live price fetch; 0 means static data only
        sourcesCount:
          type: integer
          description: Number of verified data sources contributing to this record
      required:
        - ticker
        - company
        - price
        - exchange
        - industry
        - daysUnderOne
        - daysToDeadline
        - bounceProbability
        - delistingRisk
        - complianceScore
        - financialScore
        - operatorScore
        - industryScore
        - patternScore
        - tradabilityScore
        - operatorNote
        - financialNote
        - tradeWindow
        - entryZone
        - targetZone
        - stopZone
        - status
        - volume
        - chart
        - confidenceScore
        - dataFreshnessMs
        - sourcesCount
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...