# lib/api-zod/src/generated/types/stock.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChartPoint } from "./chartPoint";

export interface Stock {
  ticker: string;
  company: string;
  price: number;
  exchange: string;
  industry: string;
  daysUnderOne: number;
  daysToDeadline: number;
  bounceProbability: number;
  delistingRisk: number;
  complianceScore: number;
  financialScore: number;
  operatorScore: number;
  industryScore: number;
  patternScore: number;
  tradabilityScore: number;
  operatorNote: string;
  financialNote: string;
  tradeWindow: string;
  entryZone: string;
  targetZone: string;
  stopZone: string;
  status: string;
  volume: string;
  chart: ChartPoint[];
  /** Model confidence 0–1 based on data freshness and source availability */
  confidenceScore: number;
  /** Unix epoch ms of the last successful live price fetch; 0 means static data only */
  dataFreshnessMs: number;
  /** Number of verified data sources contributing to this record */
  sourcesCount: number;
}